Added primitive shooting and death, and updated .gitignore
Player death is server-authoritative with clients handling things like view model hiding and resetting max health after respawn. This could be moved to server but meh. Added packets and packet handlers for health and respawning sent from the server to the client, as well as a packet and packet handlers for shoot inputs. Sets the stage for freakin sick FPS gameplay